refactor(FileActionsFloatingPanelViewController): Reduce complexity, preserving existing logic
